requirements.txt
APScheduler==3.6.3
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
idna==2.10
phabricator==0.8.0
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.4
requests==2.25.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
